chore: re-enable linting
When the JS code was moved into the js/ directory the aegir linting task was no longer considering them for linting. This is because the default globs do not include this path.

This PR adds custom lint globs to the aegir config and fixes linting errors.

depends on ipfs/aegir#240

License: MIT
Signed-off-by: Alan Shaw <alan@tableflip.io>