fix: ensure test for resolve recursive has another node
The resolve IPNS link recursively test uses name.publish to publish an IPNS link as the value for the peer. It then uses resolve to resolve the peer ID and recursively then resolve the IPNS link. This fixes the test so that when name.publish is called there is another connected node to publish to so that we do not get the error "failed to find any peer in table", in the case that the node has not yet connected to any other peers.

License: MIT
Signed-off-by: Alan Shaw <alan@tableflip.io>