refactor: relax check for undefined property
If `x` has a property explicitly set to `undefined` then `expect(x).to.eql(y)` requires `y` to also have a property set to undefined. This PR relaxes that requirement to also allow `y` to simply not have the property.

This is to save having to explicitly re-define this property to undefined when y has been `JSON.stringify`'d. This happens in the ipfs-postmsg-proxy and this change removes the need for the proxy to know the shape of this object (or at least which properties might have been undefined).

License: MIT
Signed-off-by: Alan Shaw <alan.shaw@protocol.ai>