fix: swarm addrs test
If the common factory creates a node with no bootstrap nodes then getting a list of swarm addresses could be empty (or could also be empty anyway if the connection to the bootstrap nodes takes a long time). This PR spawns and connects another IPFS node to ensure that there should be swarm addresses available.

License: MIT
Signed-off-by: Alan Shaw <alan.shaw@protocol.ai>