Added in a test for the opts parameter.
The opts parameter wasn't correctly passed in when calling ipfs.object.stat. This is the test that makes sure opts is being correctly passed to through the send function to the IPFS daemon. 

I've also removed the "base58" encoding that was being passed in for some tests as base58 isn't a valid option for encoding and would provide this error: 

Uncaught AssertionError: expected [Error: invalid encoding: base58] to not exist