fix: remove the setup of extra nodes
We previously set this node up and attempted to dial it without a
relay node being present which failed and we accidentally-on-purpose
ignored the error by resolving a promise with an error instead of
rejecting it.

The test does not assert anything to do with the other node, and the
`allowOffline` option should allow us to publish IPNS names without
having any peers so the other node is redundant and can be removed.