refactor: return peer ids as strings
As part of the async iterator refactor some peer IDs were being
returned as CIDs - the conversion was done in IPFS but not in
libp2p, which meant that where IPFS exposes libp2p directly, peer
IDs were being returned as strings, e.g. pubsub topic subscribers,
message senders, etc.

The aim is to convert these to CIDs in the long run but such a
change needs to be driven from inside libp2p instead of piecemeal
at the IPFS layer.

This PR changes the `ipfs.id().id` and `ipfs.swarm.peers().[].peer`
properties to strings so we can do the CID conversion in one go.

Also adds more in-depth tests for `ipfs.id`.